ui code cleanup
