+++ AIForecast/modeling/dataprocessing.py
@@ -392,49 +392,5 @@ class ModelEvaluationReporter:
         return pd.DataFrame(model_fit, columns=cols, index=_set.index)
 
     def save(self, file_loc: str):
-        # Todo: Make this work
-        self.train_fit.to_csv(f'{FolderStructure.CLIMATE_DATA_DIR.get_path()}\\model_training_report.csv')
-        self.test_fit.to_csv(f'{FolderStructure.CLIMATE_DATA_DIR.get_path()}\\model_testing_report.csv')
-
-
-# File Selector : csv file -> Training
-# File Selector : model schema -> Training
-# Drop down to select an imputer : None, Simple, Iterative
-# Drop down to select a split type : Straight, Rolling, Expanding
-
-# Straight split:
-    # Training percentage: a % -> 80% or 0.8 Optional
-    # Validation percentage: a % -> 0% or 0.0 Optional
-
-# Rolling Split:
-    # training size: an integer or percent -> 0.2 or 20%
-    # validation size: an integer -> 0%
-    # testing size: an integer -> 10%
-    # gap size: an integer -> 0
-    # stride: an integers -> 1
-
-# Expanding Split:
-    # initial training size: an integer or percentage
-    # validation size: an integer or percentage 0%
-    # gap size: an integer -> 0
-    # expansion rate: an integer -> 1
-
-# Drop down to select normalization type: Min-Max norm, Z-standardization
-# Training Features = [], output features = []
-# Input width: integer -> 3
-# Output width: integer -> 1
-# stride: integer -> 1
-# gap: integer -> 0
-
-# Text field for hyper-parameters <- may or may not get implemented
-
-# Button <label train model>
-# Output for a model evaluation and data plots. -> data plots will plot the predictions of the train and test data.
-# Button to save trained model.
-
-# Test menu:
-# File Selector: trained model file.
-# Time horizon: integer -> months
-# Submit button
-# Output window with a plot plotting a graph from the end of the data to the future time horizon, plotting the
-# forecast into the future.
+        self.train_fit.to_csv(f'{file_loc}_training_report.csv')
+        self.test_fit.to_csv(f'{file_loc}_testing_report.csv')
