+++ AIForecast/ui/constants.py
@@ -1,0 +1,45 @@
+
+BACKGROUND_COLOR = '#525453'
+"""
+Default background color for most UI elements.
+"""
+
+FOREGROUND_COLOR = 'white'
+
+H1_FONT = 100
+
+BUTTON_BORDER_WIDTH = 1
+BUTTON_BACKGROUND = 'white'
+BUTTON_ACTIVE_BACKGROUND = 'gray'
+BUTTON_FOREGROUND = 'black'
+
+"""Offset from the edge of the window"""
+ALIGN_X = 15
+ALIGN_Y = 15
+"""Offset used between labels and their respective elements."""
+LABEL_OFFSET = 5
+"""X Offset for each unit width of elements. 
+Used to add up each element with a specific width in a line"""
+X_30_UNIT_WIDTH_OFFSET = 30
+X_75_UNIT_WIDTH_OFFSET = 75
+X_100_UNIT_WIDTH_OFFSET = 100
+X_110_UNIT_WIDTH_OFFSET = 110
+X_150_UNIT_WIDTH_OFFSET = 150
+X_200_UNIT_WIDTH_OFFSET = 200
+"""The x and y offsets between other elements in a frame"""
+X_ELEMENT_OFFSET = 15
+Y_ELEMENT_OFFSET = 40
+"""Alignment offsets used to align certain elements with other elements vertically"""
+OUTPUT_WIDTH_ALIGNMENT_OFFSET = 20
+STRIDE_ALIGNMENT_OFFSET = 20
+TIME_OFFSET_ALIGNMENT_OFFSET = 40
+EPOCH_LABEL_ALIGNMENT_OFFSET = 50
+LEFT_BUTTON_ALIGNMENT_OFFSET = 105
+RIGHT_BUTTON_ALIGNMENT_OFFSET = 435
+
+CSV_FILE_TYPE = '*.csv'
+CSV_FILE_LABEL = 'CSV Files'
+JSON_FILE_TYPE = '*.json'
+JSON_FILE_LABEL = 'JSON Files'
+MODEL_FILE_TYPE = '*.h5'
+MODEL_FILE_LABEL = 'Keras Model Files'
